refactor: export js logic to dedicated files for better readability and maintainability
