+++ DesignPatternsWorkshop.Presentation/Views/Home/Index.cshtml
@@ -28,4 +28,4 @@
 
 
 <script src="https://cdnjs.cloudflare.com/ajax/libs/microsoft-signalr/7.0.5/signalr.min.js"></script>
-<script src="~/js/signalRConnection.js"></script>
+<script type="module" src="~/js/signalRConnection.js"></script>
