+++ DesignPatternsWorkshop.Presentation/wwwroot/js/redoLast.js
@@ -1,0 +1,5 @@
+import { connection } from "/js/signalRConnection.js"
+
+export default function redoLast() {
+  connection.invoke("Redo").catch(error => console.error(error))
+}
