+++ DesignPatternsWorkshop.Presentation/wwwroot/js/signalRConnection.js
@@ -1,19 +1,14 @@
-const connection = new signalR.HubConnectionBuilder()
+import addProduct from "/js/addProduct.js"
+import removeProduct from "/js/removeProduct.js"
+import undoLast from "/js/undoLast.js"
+import redoLast from "/js/redoLast.js"
+
+export const connection = new signalR.HubConnectionBuilder()
   .withUrl("http://localhost:5130/purchase-hub")
   .withAutomaticReconnect()
   .build()
 
 connection.start().catch(error => console.error(error))
-
-function addProduct(name, price, quantity) {
-  const product = {
-    id: Math.floor(Math.random() * 1000).toString(),
-    name: name,
-    price: price,
-    quantity: quantity,
-  }
-  connection.invoke("AddProduct", product).catch(error => console.error(error))
-}
 
 connection.on("UpdatePurchase", function () {
   fetch("Home/GetProductList")
@@ -31,23 +26,7 @@ connection.on("UpdatePurchase", function () {
     .catch(error => console.error(error))
 })
 
-function removeProduct(id, name, price, quantity) {
-  const product = {
-    id,
-    name,
-    price,
-    quantity,
-  }
-
-  connection
-    .invoke("RemoveProduct", product)
-    .catch(error => console.error(error))
-}
-
-function undoLast() {
-  connection.invoke("Undo").catch(error => console.error(error))
-}
-
-function redoLast() {
-  connection.invoke("Redo").catch(error => console.error(error))
-}
+window.addProduct = addProduct
+window.removeProduct = removeProduct
+window.undoLast = undoLast
+window.redoLast = redoLast
