+++ DesignPatternsWorkshop.Presentation/wwwroot/js/undoLast.js
@@ -1,0 +1,5 @@
+import { connection } from "/js/signalRConnection.js"
+
+export default function undoLast() {
+  connection.invoke("Undo").catch(error => console.error(error))
+}
